+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ## [1.0.0]
 
+- 1.07 Add Symbol: + public - private # protected ~ package
+  - S static O Override A abstract C Constructor o open
 - 1.06 Async for 2023.3
 - 1.05 Support Kotlin call graph
 - 1.04 PlantUML and Graphviz
@@ -14,6 +16,8 @@
 
 # 中文更新日志
 
+- 1.07 添加符号：+ 公有 - 私有 # 保护 ~ 包级
+  - S 静态 O 重写 A 抽象 C 构造 o 可重写
 - 1.06 异步用于 2023.3
 - 1.05 支持 Kotlin 调用图
 - 1.04 PlantUML 与 Graphviz 实现